+++ src/pages/Signup.js
@@ -67,13 +67,14 @@ const Signup = () => {
                 const { logradouro, uf, localidade, bairro } = response.data;
                 setZipCodeData(response.data);
 
+                console.log('cep cortado', String(digitsOnly).substring(0, 7))
                 setUser((prevUser) => ({
                     ...prevUser,
                     street: logradouro,
                     state: uf,
                     city: localidade,
                     neighborhood: bairro,
-                    zip_code: value,
+                    zip_code: String(digitsOnly).substring(0, 7),
                 }));
             } else {
                 setZipCodeData(null);
@@ -124,7 +125,7 @@ const Signup = () => {
         } else {
             numeros = [];
         }
-        return numeros[0] + numeros[1];
+        return Number(numeros[0] + numeros[1]);
     }
 
     const handleSubmit = (e) => {
@@ -134,38 +135,44 @@ const Signup = () => {
             number: user.number,
             city: user.city,
             state: user.state,
-            zip_code: extrairNumeros(user.zip_code),
+            zip_code: String(extrairNumeros(user.zip_code)),
             complement: user.complement,
             neighborhood: user.neighborhood
         }
-
-        addressService.create(address).then(response => {
-            console.log('response addres', response)
-            if (response.status === 201) {
-                user.address_id = Number(response?.data?.id_address);
-                console.log('selectedImage', selectedImage)
-                user.image = selectedImage;
-                userService.create(user).then(response => {
+        if (String(address.state).toLowerCase() === 'sp' && String(address.city).toLowerCase() === 'franca') {
+            {
+                addressService.create(address).then(response => {
+                    console.log('response addres', response)
                     if (response.status === 201) {
-                        toast.success('Usuário adicionado com sucesso!');
-                        resetForm();
+                        user.address_id = Number(response?.data?.id_address);
+                        console.log('selectedImage', selectedImage)
+                        user.image = selectedImage;
+                        userService.create(user).then(response => {
+                            if (response.status === 201) {
+                                toast.success('Usuário adicionado com sucesso!');
+                                resetForm();
+                            }
+                        }).catch(error => {
+                            toast.error(`${error.response.data.message}`);
+                            addressService.delete(user.address_id).then(response => {
+                                console.log('response delete adddres', response)
+                            })
+                            console.log(error)
+                        })
+
                     }
-                }).catch(error => {
-                    toast.error(`${error.response.data.message}`);
-                    addressService.delete(user.address_id).then(response => {
-                        console.log('response delete adddres', response)
-                    })
-                    console.log(error)
-                })
+                }).catch(
+                    error => {
+                        console.log(error);
+                        toast.error(`${error.response.data.message}`);
 
+                    }
+                )
             }
-        }).catch(
-            error => {
-                console.log(error);
-                toast.error(`${error.response.data.message}`);
+        } else {
+            toast.warning('Desculpe-nos! Por enquanto só é possível efetuar cadastro para a cidade de Franca-SP');
+        }
 
-            }
-        )
     };
 
     const cpfMask = [
